feat(spring-core): add simple aspect example
