--- 05-spring-core/src/main/java/com/github/alebabai/jmp2k17/service/EventsFactory.java
@@ -1,7 +1,0 @@
-package com.github.alebabai.jmp2k17.service;
-
-import com.github.alebabai.jmp2k17.domain.Event;
-
-public interface EventsFactory {
-    Event createEvent();
-}
