--- 05-spring-core/src/main/java/com/github/alebabai/jmp2k17/service/EventsInvokerService.java
@@ -1,7 +1,0 @@
-package com.github.alebabai.jmp2k17.service;
-
-import com.github.alebabai.jmp2k17.domain.Event;
-
-public interface EventsInvokerService {
-    void invoke(Event event);
-}
