+++ 05-spring-core/src/main/java/com/github/alebabai/jmp2k17/spring/service/EventsBroadcasterService.java
@@ -1,4 +1,4 @@
-package com.github.alebabai.jmp2k17.service;
+package com.github.alebabai.jmp2k17.spring.service;
 
 public interface EventsBroadcasterService {
     void start();
