+++ 05-spring-core/src/main/java/com/github/alebabai/jmp2k17/spring/service/EventsInvokerService.java
@@ -1,0 +1,7 @@
+package com.github.alebabai.jmp2k17.spring.service;
+
+import com.github.alebabai.jmp2k17.spring.domain.Event;
+
+public interface EventsInvokerService {
+    void invoke(Event event);
+}
