+++ 05-spring-core/src/main/java/com/github/alebabai/jmp2k17/spring/service/impl/EventsInvokerServiceImpl.java
@@ -1,0 +1,12 @@
+package com.github.alebabai.jmp2k17.spring.service.impl;
+
+import com.github.alebabai.jmp2k17.spring.domain.Event;
+import com.github.alebabai.jmp2k17.spring.service.EventsInvokerService;
+
+public class EventsInvokerServiceImpl implements EventsInvokerService {
+
+    @Override
+    public void invoke(Event event) {
+        System.out.println(event);
+    }
+}
